--- commons/pdf-commons/src/main/java/se/idsec/signservice/pdf/package-info.java
@@ -1,4 +1,0 @@
-/**
- * PDF support.
- */
-package se.idsec.signservice.pdf;
