--- commons/pdf-commons/src/main/java/se/idsec/signservice/pdf/sign/package-info.java
@@ -1,4 +1,0 @@
-/**
- * PDF Signature classes
- */
-package se.idsec.signservice.pdf.sign;
