+++ commons/pdf-commons/src/main/java/se/idsec/signservice/security/sign/pdf/PDFSignerResult.java
@@ -15,8 +15,7 @@
  */
 package se.idsec.signservice.security.sign.pdf;
 
-import org.bouncycastle.cms.CMSSignedData;
-import se.idsec.signservice.pdf.sign.PDFSignTaskDocument;
+import se.idsec.signservice.security.sign.pdf.document.PDFSignTaskDocument;
 import se.idsec.signservice.security.sign.SignerResult;
 
 import java.security.cert.X509Certificate;
