+++ commons/pdf-commons/src/main/java/se/idsec/signservice/security/sign/pdf/configuration/package-info.java
@@ -1,0 +1,1 @@
+package se.idsec.signservice.security.sign.pdf.configuration;
