+++ commons/pdf-commons/src/main/java/se/idsec/signservice/security/sign/pdf/document/PDFSignTaskDocument.java
@@ -1,4 +1,4 @@
-package se.idsec.signservice.pdf.sign;
+package se.idsec.signservice.security.sign.pdf.document;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
