+++ commons/pdf-commons/src/main/java/se/idsec/signservice/security/sign/pdf/signprocess/CMSProcessableInputStream.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package se.idsec.signservice.pdf.general;
+package se.idsec.signservice.security.sign.pdf.signprocess;
 
 import org.bouncycastle.asn1.ASN1ObjectIdentifier;
 import org.bouncycastle.asn1.cms.CMSObjectIdentifiers;
