+++ commons/pdf-commons/src/main/java/se/idsec/signservice/security/sign/pdf/signprocess/SubjectDnAttribute.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package se.idsec.signservice.pdf.general;
+package se.idsec.signservice.security.sign.pdf.signprocess;
 
 /**
  *
