+++ commons/pdf-commons/src/main/java/se/idsec/signservice/security/sign/pdf/signprocess/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * PDF Signature process functions
+ */
+package se.idsec.signservice.security.sign.pdf.signprocess;
